Resolved code smells

+++ data-retrieval/taxonomy-fetcher/src/impls/open_tree_of_life/taxonomy.rs
@@ -12,7 +12,7 @@ pub struct OpenTreeOfLifeTaxonomy {
     /// Version of the Open Tree of Life taxonomy.
     pub version: OpenTreeOfLifeVersion,
     /// Root of the taxonomy.
-    pub root_position: u32,
+    pub root_position: usize,
     /// Taxon entries.
     pub taxon_entries: Vec<OpenTreeOfLifeTaxonEntry>,
 }
